--- a/sllm_store/Dockerfile.builder
+++ b/sllm_store/Dockerfile.builder
@@ -62,7 +62,6 @@ COPY sllm_store ./sllm_store
 COPY setup.py .
 COPY pyproject.toml .
 COPY MANIFEST.in .
-COPY setup.cfg .
 COPY requirements.txt .
 COPY README.md .
 COPY proto ./proto