--- a/sllm_store/Dockerfile.rocm
+++ b/sllm_store/Dockerfile.rocm
@@ -38,7 +38,6 @@ COPY sllm_store ./sllm_store
 COPY setup.py .
 COPY pyproject.toml .
 COPY MANIFEST.in .
-COPY setup.cfg .
 COPY requirements.txt .
 COPY README.md .
 COPY proto ./proto