--- a/sllm_store/cmake/hipify.py
+++ b/sllm_store/cmake/hipify.py
@@ -14,9 +14,30 @@ import shutil
 
 from torch.utils.hipify.hipify_python import hipify
 
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser()
 
-def vllm_hipify(args):
-    # copied from vllm's cmake/hipify.py
+    # Project directory where all the source + include files live.
+    parser.add_argument(
+        "-p",
+        "--project_dir",
+        help="The project directory.",
+    )
+
+    # Directory where hipified files are written.
+    parser.add_argument(
+        "-o",
+        "--output_dir",
+        help="The output directory.",
+    )
+
+    # Source files to convert.
+    parser.add_argument(
+        "sources", help="Source files to hipify.", nargs="*", default=[]
+    )
+
+    args = parser.parse_args()
+
     # Limit include scope to project_dir only
     includes = [os.path.join(args.project_dir, "*")]
 
@@ -55,57 +76,3 @@ def vllm_hipify(args):
 
     # Print hipified source files.
     print("\n".join(hipified_sources))
-
-
-def sllm_hipify(args):
-    extra_files = [os.path.abspath(s) for s in args.sources]
-    temp_dir = os.path.join(args.project_dir, "temp")
-    hipify_result = hipify(
-        project_directory=os.path.abspath(args.project_dir),
-        output_directory=os.path.abspath(temp_dir),
-        show_detailed=True,
-        is_pytorch_extension=True,
-        extra_files=extra_files,
-    )
-    os.makedirs(args.output_dir, exist_ok=True)
-    for _, result in hipify_result.items():
-        directory, file_name = os.path.split(result.hipified_path)
-        shutil.copy(
-            result.hipified_path, os.path.join(args.output_dir, file_name)
-        )
-
-    shutil.rmtree(temp_dir)
-
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-
-    # Project directory where all the source + include files live.
-    parser.add_argument(
-        "-p",
-        "--project_dir",
-        help="The project directory.",
-    )
-
-    # Directory where hipified files are written.
-    parser.add_argument(
-        "-o",
-        "--output_dir",
-        help="The output directory.",
-    )
-
-    # Source files to convert.
-    parser.add_argument(
-        "sources", help="Source files to hipify.", nargs="*", default=[]
-    )
-
-    # vLLM mode or not
-    parser.add_argument(
-        "-s", "--sllm", help="hipify for sllm", action="store_true"
-    )
-
-    args = parser.parse_args()
-    if args.sllm:
-        sllm_hipify(args)
-    else:
-        vllm_hipify(args)