--- a/sllm_store/csrc/sllm_store/checkpoint_store.cu
+++ b/sllm_store/csrc/sllm_store/checkpoint_store.cu
@@ -15,8 +15,6 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 //  ----------------------------------------------------------------------------
-#include "checkpoint_store.h"
-
 #include <fcntl.h>
 #include <glog/logging.h>
 #include <sys/stat.h>
@@ -27,6 +25,7 @@
 #include <filesystem>
 #include <thread>
 
+#include "checkpoint_store.h"
 #include "error_handling.h"
 
 CheckpointStore::CheckpointStore(const std::string& storage_path,