--- a/sllm_store/requirements-build-rocm.txt
+++ b/sllm_store/requirements-build-rocm.txt
@@ -1,6 +1,4 @@
 cmake>=3.20
-grpcio==1.49.1
-grpcio-tools==1.49.1
 ninja
 setuptools
 wheel