--- a/sllm_store/setup.cfg
+++ b/sllm_store/setup.cfg
@@ -1,2 +1,0 @@
-[options.data_files]
-. = requirements.txt