### Description

This PR replaces the C++ implementation with a Python-based implementation, primarily to reduce the long build times of gRPC/C++ during development. Key benefits include:

1. **Reduced Build Time:**  
   - **New CMake compile time:** 69s  
   - **Old CMake compile time:** 197s  
   - **New `pip install` time:** 180s  
   - **Old `pip install` time:** 362s  

2. **Significant CPU Savings:**  
   - **New `time` output:** ~192% CPU  
   - **Old `time` output:** ~1182% CPU  
   - ~6x CPU usage reduction, especially beneficial for local development on lower-end hardware.

3. **Enhanced Usability for CLI:**  
   - Introduced `click` for CLI, which simplifies handling command-line inputs.  
   - Memory size parameters now accept strings (e.g., `--memory 1GB` instead of `--memory 1024`) for more flexible input.

### Motivation

This PR addresses performance bottlenecks caused by gRPC/C++ build times, simplifying local development. It closes #161.

### Type of Change

- [ ] Bug fix  
- [x] New feature  
- [x] Breaking change  
- [x] Documentation update  

### Checklist

- [x] I have read the [CONTRIBUTION](https://github.com/ServerlessLLM/ServerlessLLM/blob/main/CONTRIBUTING.md) guide.  
- [ ] I have updated the tests (if applicable).  
- [x] I have updated the documentation (if applicable).