@SiyangShao 2025-01-17
I've tried the new build target. In order to trigger hipify automatically, here's what we need to do:

1. We have to manually change the `.cpp` files related to GPU operations under `sllm_store/csrc` into `.cu` files.
2. We need to manually set `CHECKPOINT_STORE_SOURCES`, compared with using `GLOB`

In summary, we need to change the filename and `CMakeLists.txt` into following:
```
set(CHECKPOINT_STORE_SOURCES
  "csrc/sllm_store/binary_utils.cpp"
  "csrc/sllm_store/checkpoint_store.cu"
  "csrc/sllm_store/checkpoint_store_py.cu"
  "csrc/sllm_store/cuda_memory.cu"
  "csrc/sllm_store/cuda_memory_pool.cu"
  "csrc/sllm_store/gpu_replica.cu"
  "csrc/sllm_store/memory_state.cpp"
  "csrc/sllm_store/model.cu"
  "csrc/sllm_store/pinned_memory.cpp"
  "csrc/sllm_store/pinned_memory_pool.cu"
)
```
In addition, we can change the `hipfiy.py` under `cmake` folder same as [vLLM](https://github.com/vllm-project/vllm/blob/main/cmake/hipify.py). (Current version includes our own version of hipify script, which is no longer needed.)

After that, we shall find the following `hipify` log during compilation:
```
[10/21] cd /app/build/temp.linux-x86_64-cpython-310 && /app/cmake/hipify.py -p /app/csrc -o /app/build/temp.linux-x86_64-cpython-310/csrc csrc/sllm_store/checkpoint_store.cu csrc/sllm_store/checkpoint_store_py.cu csrc/sllm_store/cuda_memory.cu csrc/sllm_store/cuda_memory_pool.cu csrc/sllm_store/gpu_replica.cu csrc/sllm_store/model.cu csrc/sllm_store/pinned_memory_pool.cu
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/cuda_memory_pool.h -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/hip_memory_pool.h [ok]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/cuda_memory.h -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/hip_memory.h [ok]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/error_handling.h -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/error_handling_hip.h [ok]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/pinned_memory_pool.h -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/pinned_memory_pool.h [skipped, no changes]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/pinned_memory.h -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/pinned_memory.h [skipped, no changes]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/concurrent_queue.h -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/concurrent_queue.h [skipped, no changes]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/concurrent_vector.h -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/concurrent_vector.h [skipped, no changes]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/memory_state.h -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/memory_state.h [skipped, no changes]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/types_and_defs.h -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/types_and_defs.h [skipped, no changes]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/model.h -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/model_hip.h [ok]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/checkpoint_store.h -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/checkpoint_store_hip.h [ok]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/checkpoint_store.cu -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/checkpoint_store.hip [ok]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/checkpoint_store_py.cu -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/checkpoint_store_py.hip [ok]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/cuda_memory.cu -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/hip_memory.hip [ok]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/cuda_memory_pool.cu -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/hip_memory_pool.hip [ok]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/gpu_replica.h -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/gpu_replica.h [skipped, no changes]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/gpu_replica.cu -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/gpu_replica.hip [ok]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/model.cu -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/model.hip [ok]
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/pinned_memory_pool.cu -> /app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/pinned_memory_pool.hip [ok]
Successfully preprocessed all matching files.
Total number of unsupported CUDA function calls: 0


Total number of replaced kernel launches: 0
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/checkpoint_store.hip
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/checkpoint_store_py.hip
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/hip_memory.hip
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/hip_memory_pool.hip
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/gpu_replica.hip
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/model.hip
/app/build/temp.linux-x86_64-cpython-310/csrc/sllm_store/pinned_memory_pool.hip
```