@SiyangShao 2025-01-17
I've tried to install the latest `sllm-store` via `pip install .`. Save model works fine, but for server, I faced such an issue:
```
❯ sllm-store --help
Traceback (most recent call last):
  File "/home/shaosy/miniconda3/envs/sllm_load/bin/sllm-store", line 5, in <module>
    from sllm_store.cli import main
  File "/home/shaosy/miniconda3/envs/sllm_load/lib/python3.10/site-packages/sllm_store/cli.py", line 6, in <module>
    from sllm_store.server import serve
  File "/home/shaosy/miniconda3/envs/sllm_load/lib/python3.10/site-packages/sllm_store/server.py", line 15, in <module>
    from sllm_store._checkpoint_store import (  # noqa: E402
ImportError: /home/shaosy/miniconda3/envs/sllm_load/bin/../lib/libstdc++.so.6: version `GLIBCXX_3.4.30' not found (required by /home/shaosy/miniconda3/envs/sllm_load/lib/python3.10/site-packages/sllm_store/_checkp
oint_store.cpython-310-x86_64-linux-gnu.so)
```